Moved proper stuff from UserAuthAPI to UserDB in effort to start removing all backend code
